Move container into new inventory-package
